Start new simpler structure and migration from the FE wiki
