--- admin/README.md
@@ -1,13 +1,0 @@
-# Admin
-
-Various helpers for administrating a DSS instance.
-
-## Jobs
-
-* `list_jobs_by_status`: list jobs by current status in a given project
-* `filter_jobs_by_start_date`: list jobs that were started after a specific date
-* `abort_all_running_jobs`: terminate all running jobs in a project
-
-## Projects
-
-* `edit_project_permissions` : grant or revoke permissions for a given group
