+++ admin/edit_project_permissions/README.md
@@ -1,0 +1,7 @@
+# Edit project permissions
+
+## Questions
+* I want to programmatically add/remove group permissions for a given project.
+
+## Answers
+This is possible with the `set_permissions()` method for [projects](https://doc.dataiku.com/dss/latest/python-api/projects.html#modifying-project-settings) in the DSS public API.
