+++ admin/list_and_filter_jobs/README.md
@@ -1,0 +1,9 @@
+# List and filter jobs
+
+## Questions
+
+* I want to list jobs and their respective status at a given time.
+* I want to list jobs that were started after a specific date.
+
+## Answers
+This is possible using the [`list_jobs()`](https://doc.dataiku.com/dss/latest/python-api/jobs.html) method from the DSS public API.
